style: Use /* */ comment separators consistently
As indicated in CODING_STYLE.rst, qemu code should use /* and */ for comments.

    Comment style
    =============

    We use traditional C-style /``*`` ``*``/ comments and avoid // comments.

    Rationale: The // form is valid in C99, so this is purely a matter of
    consistency of style. The checkpatch script will warn you about this.

This was not consistently done in the past.

This series is not marked as trivial because, while most of the changes were
automated, simple regexp matching had trouble not changing strings like http://
that may happen in the code and not just in comments. Some cases had to be
corrected manually.

Signed-off-by: Christophe de Dinechin <dinechin@redhat.com>